+++ src/voronoi_simple.c
@@ -7,7 +7,7 @@
 #include "common.h"
 
 static Color *pixel_buf = 0;
-static size_t buf_capacity = 0;
+static u64 buf_capacity = 0;
 
 float dist_sqr(float x1, float y1, float x2, float y2) {
     return (x1-x2)*(x1-x2) + (y1-y2)*(y1-y2);
@@ -24,8 +24,8 @@ void finish_voronoi(void) {
 
 
 void draw_voronoi(RenderTexture2D target, Vector2 *points, Color *colors, size_t num_points) {
-    size_t width  = target.texture.width;
-    size_t height = target.texture.height;
+    u64 width  = target.texture.width;
+    u64 height = target.texture.height;
 
     if (buf_capacity < width * height) {
         buf_capacity = width * height;
@@ -36,13 +36,13 @@ void draw_voronoi(RenderTexture2D target, Vector2 *points, Color *colors, size_t
 
     PROFILER_ZONE("Calculate pixel buffer");
 
-        for (size_t j = 0; j < height; j++) {
-            for (size_t i = 0; i < width; i++) {
+        for (u64 j = 0; j < height; j++) {
+            for (u64 i = 0; i < width; i++) {
 
                 // find the closest point
-                size_t close_index = 0;
+                u64 close_index = 0;
                 float d1 = dist_sqr(points[0].x, points[0].y, i, j);
-                for (size_t k = 1; k < num_points; k++) {
+                for (u64 k = 1; k < num_points; k++) {
                     float d2 = dist_sqr(points[k].x, points[k].y, i, j);
                     if (d2 < d1) {
                         d1 = d2;
@@ -60,12 +60,12 @@ void draw_voronoi(RenderTexture2D target, Vector2 *points, Color *colors, size_t
     PROFILER_ZONE("draw into texture");
     BeginTextureMode(target);
 
-    for (size_t j = 0; j < height; j++) {
+    for (u64 j = 0; j < height; j++) {
         // find a band of the same color.
         // this is MUCH faster than just useing DrawPixel()
-        size_t i = 0;
+        u64 i = 0;
         while (i < width) {
-            size_t low_i = i;
+            u64 low_i = i;
             Color this_color = pixel_buf[j*width + i];
             for (; i < width; i++) {
                 if (!ColorIsEqual(this_color, pixel_buf[j*width + i])) break;
